# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StandupInput {
  /**
     * Raw messy notes from the workday
     * @minLength 1
     */
  notes: string;
}

export interface StandupOutput {
  /** What was done yesterday */
  yesterday: string;
  /** What is planned for today */
  today: string;
  /** Any blockers or impediments */
  blockers: string;
}

export interface ErrorResponse {
  error: string;
}