# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: standup
    description: Standup generation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /standup/generate:
    post:
      operationId: generateStandup
      tags: [standup]
      summary: Generate standup update
      description: Takes messy notes and generates a structured standup update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StandupInput"
      responses:
        "200":
          description: Generated standup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandupOutput"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    StandupInput:
      type: object
      properties:
        notes:
          type: string
          description: Raw messy notes from the workday
          minLength: 1
      required:
        - notes
    StandupOutput:
      type: object
      properties:
        yesterday:
          type: string
          description: What was done yesterday
        today:
          type: string
          description: What is planned for today
        blockers:
          type: string
          description: Any blockers or impediments
      required:
        - yesterday
        - today
        - blockers
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error