# lib/api-zod/src/generated/types/standupInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StandupInput {
  /**
     * Raw messy notes from the workday
     * @minLength 1
     */
  notes: string;
}