# lib/api-zod/src/generated/types/standupOutput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StandupOutput {
  /** What was done yesterday */
  yesterday: string;
  /** What is planned for today */
  today: string;
  /** Any blockers or impediments */
  blockers: string;
}